created superclass and renamed TuelibMixin to TuelibBiblioMixin, deleted various unused vars
