+++ java/solr_plugins/RangedSearches/src/de/uni_tuebingen/ub/ixTheo/bibleRangeSearch/BibleRangeParser.java
@@ -14,7 +14,6 @@ import org.apache.solr.search.SyntaxError;
 public class BibleRangeParser extends QParser {
     private final static String QUERY_SEPARATOR = " ";
     private final static String DB_FIELD_SEPARATOR = ",";
-    //private static Logger logger = LoggerFactory.getLogger(BibleRangeParser.class);
 
     /**
      * Constructor for the QParser
