+++ java/solr_plugins/RangedSearches/src/de/uni_tuebingen/ub/ixTheo/bibleRangeSearch/BibleRangeWeight.java
@@ -24,7 +24,6 @@ import de.uni_tuebingen.ub.ixTheo.rangeSearch.RangeScorer;
 
 public class BibleRangeWeight extends ConstantScoreWeight {
     private final static float NOT_RELEVANT = Float.NEGATIVE_INFINITY;
-    //private final static Logger logger = LoggerFactory.getLogger(BibleRangeWeight.class);
     private final static String FIELD = "bible_ranges";
     private final IntFloatMap scoring = new IntFloatHashMap();
     private final boolean isSearchingForBooks;
