+++ java/solr_plugins/RangedSearches/src/de/uni_tuebingen/ub/ixTheo/timeAspectDateRangeSearch/TimeAspectDateRangeParser.java
@@ -9,7 +9,6 @@ import org.apache.solr.search.SyntaxError;
 
 
 public class TimeAspectDateRangeParser extends QParser {
-    //private static Logger logger = LoggerFactory.getLogger(TimeAspectDateRangeParser.class);
 
     /**
      * @param qstr        The part of the query string specific to this parser
