+++ java/solrmarc_mixin/src/de/unituebingen/ub/ubtools/solrmarcMixin/KrimDok.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.marc4j.marc.Record;
 import org.marc4j.marc.VariableField;
 
-public class KrimDok extends TuelibMixin {
+public class KrimDok extends TuelibBiblioMixin {
     public String isAvailableForPDA(final Record record) {
         final List<VariableField> fields = record.getVariableFields("PDA");
         return Boolean.toString(!fields.isEmpty());
